requirements.txt
absl-py==1.1.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argilla==1.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==5.0.1
attrs==23.1.0
automium==0.2.6
automium-web==0.1.1
autoreload==0.1.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
confection==0.0.4
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
DBUtils==3.0.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.5.1
distlib==0.3.9
Django==4.2.18
django-extensions==3.2.1
django-jazzmin==3.0.1
django-mysql==4.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.3.0
docopt==0.6.2
effdet==0.4.1
email-validator==2.0.0.post2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl#sha256=8902305d2ced83d98a8e88efd93ea8970a70dda9bb24b0024a8b798d1cc913d5
et-xmlfile==1.1.0
etils==0.6.0
exceptiongroup==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.0
fastjsonschema==2.17.1
filelock==3.17.0
filetype==1.2.0
Flask==1.1.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
flatbuffers==1.12
fonttools==4.33.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
googletrans==4.0.0rc1
greenlet==2.0.2
grpcio==1.46.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.14.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.7.1
iopath==0.1.10
ipykernel==6.23.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==3.1.5
joblib==1.2.0
jose==1.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langchain==0.0.191
langcodes==3.3.0
layoutparser==0.3.4
legacy==0.1.6
libclang==14.0.1
limits==4.0.1
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==3.0.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.16.5
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.22.4
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.7
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathy==0.10.1
patsy==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone-client==2.2.2
pipreqs==0.5.0
platformdirs==4.3.6
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.17.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.8
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.7.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.9.1
pyperclip==1.8.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytweening==1.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==5.2.1
regex==2023.5.5
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rsa==4.8
safetensors==0.3.1
scikit-learn==1.6.1
scipy==1.11.0
selenium==4.8.2
Send2Trash==1.8.2
six==1.16.0
slowapi==0.1.9
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.3
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.4
SQLAlchemy==2.0.15
sqlparse==0.4.2
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.4
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.5.0
tiktoken==0.4.0
timm==0.9.2
tinycss2==1.1.1
tokenizer==3.4.2
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.9.2
truecase==0.0.14
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.0
tzdata==2022.1
unstructured==0.7.9
unstructured-inference==0.5.1
uri-template==1.2.0
urllib3==1.25.11
uvicorn==0.23.1
virtualenv==20.29.1
waitress==2.1.2
Wand==0.6.11
wasabi==1.1.1
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==10.3
Werkzeug==1.0.1
widgetsnbextension==4.0.7
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarg==0.1.9
yarl==1.9.2
zeep==4.2.1
zipp==3.15.0
